user photos showing in FlatList on index.
I'm going to finish out showing the users in a flatList. Maybe I'll move it over to a different tab. But I want the first page to maybe open on a wipe layout, and if users want, they can switch to grid layout, where they  can swipe on each person.'