Swiper set up with centered image
I had to add a paddingRight to get the image centered on the screen, which is bad because I could not figure out why it would not center. I'm not sure if this will be okay, but I did this fix by viewportwidth, so we'll see. Next, I want to see why image quality is so low, and see if I can fix that, at least for aesthetic reasons while building.